+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [0.4.24] - Unreleased
+### Added
+- TilesRenderer: Added support for reverse depth buffer.
+
 ### Changed
 - EnvironmentControls: PivotMesh is now added to the scene before the "start" event is fired.
 - ImageOverlayPlugin: Do not allocate a render target for compositing if only a single tile is returned that exactly matches the tile bounds.
@@ -14,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - ImageOverlayPlugin: Fix case where UVs can be generated as NaN.
 - ImageFormatPlugin: Adjust root tile geometricError to be Infinity.
 - ImageOverlays: Updated types.
+- EnvironmentControls: Removed use of "THREE.Clock"
 
 ## [0.4.23] - 2026-03-18
 ### Added